+++ packages/browser-injected-sdk/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @phantom/browser-sdk
 
+## 1.0.5
+
+### Patch Changes
+
+- a8287d6: Fix release for auth 2
+- Updated dependencies [a8287d6]
+  - @phantom/chain-interfaces@1.0.5
+  - @phantom/constants@1.0.5
+  - @phantom/sdk-types@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes