+++ packages/constants/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @phantom/constants
 
+## 1.0.5
+
+### Patch Changes
+
+- a8287d6: Fix release for auth 2
+- Updated dependencies [a8287d6]
+  - @phantom/sdk-types@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes