+++ packages/react-sdk/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @phantom/react-sdk
 
+## 1.0.5
+
+### Patch Changes
+
+- a8287d6: Fix release for auth 2
+- Updated dependencies [a8287d6]
+  - @phantom/browser-sdk@1.0.5
+  - @phantom/chain-interfaces@1.0.5
+  - @phantom/constants@1.0.5
+  - @phantom/wallet-sdk-ui@1.0.5
+
 ## 1.0.4
 
 ### Patch Changes